made message generation separate
