+++ src/commands/mod.rs
@@ -2,3 +2,4 @@ pub mod dia;
 pub mod math;
 pub mod meta;
 pub mod owner;
+pub mod message_generation;
