+++ src/main.rs
@@ -23,7 +23,7 @@ use serenity::{
     prelude::*,
 };
 
-use commands::{dia::*, math::*, meta::*, owner::*};
+use commands::{dia::*, math::*, meta::*, owner::*, message_generation::*};
 use serenity::model::guild::{Guild, Member};
 use serenity::model::user::User;
 
@@ -42,48 +42,6 @@ const CHANNEL_LIST: [&str; 4] = [
     "535663746555707394",
     "675993960178647040",
     "569586147160883241",
-];
-
-const D_LIST: [&str; 4] = [
-    "d",
-    "D",
-    "ɖ",
-    "ɗ",
-];
-
-const I_LIST: [&str; 4] = [
-    "i",
-    "I",
-    "¡",
-    "ϊ",
-];
-
-const A_LIST: [&str; 4] = [
-    "a",
-    "A",
-    "∆",
-    "@",
-];
-
-const S_LIST: [&str; 4] = [
-    "s",
-    "S",
-    "$",
-    "$",
-];
-
-const K_LIST: [&str; 4] = [
-    "k",
-    "K",
-    "ꝅ",
-    "ꝅ",
-];
-
-const QMARK_LIST: [&str; 4] = [
-    "⁉",
-    "⁇",
-    "‽",
-    "‽",
 ];
 
 pub struct MessageInfo {
@@ -221,34 +179,4 @@ fn main() {
     if let Err(why) = client.start() {
         error!("Client error: {:?}", why);
     }
-}
-
-fn get_dia_string() -> String {
-    let mut s = String::from("");
-    let mut i = 0;
-    while i < 10 {
-        if i == 0 || i == 2 {
-            s.push_str(D_LIST[rand::thread_rng().gen_range(0, 4)]);
-        }
-        if i == 1 || i == 4 {
-            s.push_str(I_LIST[rand::thread_rng().gen_range(0, 4)]);
-        }
-        if i == 3 || i == 5 {
-            s.push_str(" ");
-        }
-        if i == 6 {
-            s.push_str(A_LIST[rand::thread_rng().gen_range(0, 4)]);
-        }
-        if i == 7 {
-            s.push_str(S_LIST[rand::thread_rng().gen_range(0, 4)]);
-        }
-        if i == 8 {
-            s.push_str(K_LIST[rand::thread_rng().gen_range(0, 4)]);
-        }
-        if i == 9 {
-            s.push_str(QMARK_LIST[rand::thread_rng().gen_range(0, 4)]);
-        }
-        i+= 1;
-    }
-    return s;
 }
